# pkg/trace/api: experiment with trace decoding