… pkg/trace/api/api_direct_decode.go