### What does this PR do?

This PR contains changes that experiment with trace decoding performance improvements.